Digital_Replica/requirements.txt
flask
pymongo
paho-mqtt
pyyaml
cryptography

